% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sample.R
\name{apply_methods}
\alias{apply_methods}
\title{Sample a data frame using different methods}
\usage{
apply_methods(data, nSamp, iter, method = NULL)
}
\arguments{
\item{data}{A data frame to be sampled.}

\item{nSamp}{The number of samples to generate.}

\item{iter}{The number of iterations to perform.}

\item{method}{The method to use for sampling. Can be "lhs" for Latin Hypercube Sampling,
"srs" for Simple Random Sampling, or "lpm" else for Balanced Sampling.}
}
\value{
A data frame containing the sampled data using each of the three methods.
}
\description{
This function takes a data frame and performs sampling using three different methods:
latin hypercube sampling (LHS), simple random sampling (SRS),
and Latin pseudo-maximum sampling (LPM).
For each method, it applies the \code{\link{apply_sample()}} function
to the data frame, with the specified number of samples and iterations.
}
\examples{
# Generate some random data
set.seed(123)
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Sample the data using utils_sample()
sampled_data <- apply_methods(data, nSamp = 10, iter = 5)

}
