% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{boot_mean}
\alias{boot_mean}
\title{Bootstrapped difference from population mean}
\usage{
boot_mean(x, i, pop)
}
\arguments{
\item{x}{A vector of numerical values.}

\item{i}{A vector of indices specifying the current bootstrap sample.}

\item{pop}{The population mean to compare to.}
}
\value{
The difference between the mean of the current bootstrap sample and the population mean.
}
\description{
This function computes the difference between the mean of a given sample and the population mean using bootstrapping.
}
