% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-boot.R
\name{boot_median}
\alias{boot_median}
\title{Bootstrapped median}
\usage{
boot_median(x, i)
}
\arguments{
\item{x}{A vector of numerical values.}

\item{i}{A vector of indices specifying the current bootstrap sample.}
}
\value{
The bootstrapped median of a sample.
}
\description{
This function computes the bootstrapped median of a given sample.
}
\keyword{internal}
