% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_plot.R
\name{bootstrap_plot}
\alias{bootstrap_plot}
\title{Plot bootstrap statistics}
\usage{
bootstrap_plot(data, statistics = NULL, bootstat = NULL, type = "mean", ...)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{statistics}{An optional vector of statistics to include in the plot.}

\item{bootstat}{An optional vector of bootstrapped statistics to include in the plot.}

\item{type}{The type of plot to create, either "mean" (default) or "box".}

\item{...}{Additional arguments passed to ggplot2.}
}
\value{
A ggplot2 object representing the plot of summary statistics.
}
\description{
This function creates a plot of bootstrapped statistics using ggplot2.
}
