% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootstraps}
\alias{bootstraps}
\title{Bootstraps of samples}
\format{
\code{sf} object.
}
\usage{
bootstraps
}
\description{
This is a nested \code{tibble} of bootstraps of statistics for \code{stats}
}
\keyword{datasets}
