% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gedi}
\alias{gedi}
\title{GEDI data}
\format{
\code{sf} object.
}
\source{
This data was collected by the GEDI mission
}
\usage{
gedi
}
\description{
This is an \code{sf} object containing the spatial data for 1000 GEDI samples in a study area.
}
\keyword{datasets}
