% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_samples.R
\name{iterate_samples}
\alias{iterate_samples}
\title{Iterate over multiple sampling methods and sample sizes}
\usage{
iterate_samples(data, metrics = NULL, nSamp, iter)
}
\arguments{
\item{data}{an \code{sf} of type POINT or \code{data.frame} object}

\item{metrics}{a vector of character strings that match desired column names in \code{data}}

\item{nSamp}{a numeric scalar greater than 0 and less than the number of rows in \code{data}}

\item{iter}{a numeric scalar greater than 0}
}
\value{
a list of sampled datasets
}
\description{
Iterate over multiple sampling methods and sample sizes
}
\details{
This function iterates over multiple sampling methods and sample sizes using \code{utils_sample()}.
The function assumes that users have removed coordinate columns if a \code{data.frame} has been provided.
Any NA values in \code{data} will be dropped.
}
\examples{
# Iterate over multiple sampling methods and sample sizes for the 'meuse' dataset
data("plots")
iterate_samples <- iterate_samples(data = plots, nSamp = 100, iter = 3)

}
