% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo}
\alias{monte_carlo}
\title{Iterate over multiple sampling methods and sample sizes}
\usage{
monte_carlo(data, metrics = NULL, nSamp, iter, method = NULL, cores = NULL)
}
\arguments{
\item{data}{an \code{sf} of type POINT or \code{data.frame} object}

\item{metrics}{a vector of character strings that match desired column names in \code{data}}

\item{nSamp}{a numeric scalar greater than 0 and less than the number of rows in \code{data}}

\item{iter}{a numeric scalar greater than 0}

\item{method}{The method to use for sampling. Can be "lhs" for Latin Hypercube Sampling,
"srs" for Simple Random Sampling, or "lpm" else for Balanced Sampling.}

\item{cores}{Number of cores to use for parallel computation}
}
\value{
a list of sampled datasets
}
\description{
Iterate over multiple sampling methods and sample sizes
}
\details{
This function iterates over multiple sampling methods and sample sizes using \code{utils_sample()}.
The function assumes that users have removed coordinate columns if a \code{data.frame} has been provided.
Any NA values in \code{data} will be dropped.
}
