% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_ahels.R
\name{monte_carlo_ahels}
\alias{monte_carlo_ahels}
\title{This function performs a Monte Carlo simulation of the ahels
method.}
\usage{
monte_carlo_ahels(
  existing,
  mraster,
  nFrac = 0.1,
  matrices = NULL,
  cores = NULL,
  ...
)
}
\arguments{
\item{existing}{A data frame of existing samples}

\item{mraster}{A raster object acting as the population}

\item{nFrac}{Scalar to use to define the number of samples to add using ahels}

\item{matrices}{Population maxtrix}

\item{cores}{The number of cores to use for the parallel
calculation.}

\item{...}{Further arguments passed to the \code{\link{ahelsmethod}}
function.}
}
\value{
A nested dataframe with the ahels samples.
}
\description{
This function performs a Monte Carlo simulation of the ahels
method.
}
