% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{plots}
\alias{plots}
\title{Plots data}
\format{
\code{sf} object.
}
\source{
This data was collected during a study of forest plots.
}
\usage{
plots
}
\description{
This is an \code{sf} object containing the spatial data for plots in a study area.
}
\keyword{datasets}
