% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sample.R
\name{sample_balanced}
\alias{sample_balanced}
\title{Balanced sampling}
\usage{
sample_balanced(data, nSamp, p = NULL)
}
\arguments{
\item{data}{A data frame to be sampled.}

\item{nSamp}{The number of samples to generate.}

\item{p}{probability string}
}
\value{
A data frame containing the sampled data.
}
\description{
This function performs Latin Point Mass Sampling (LPM) on data.
}
