% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bootstrap.R
\name{sample_bootstrap}
\alias{sample_bootstrap}
\title{Sample bootstrap function}
\usage{
sample_bootstrap(data, population, cores = NULL, R = 10000)
}
\arguments{
\item{data}{A nested dataframe with a nested column called statistics.}

\item{population}{A vector representing the population from which the bootstrap samples will be drawn.}

\item{cores}{An optional argument indicating the number of CPU cores to use for parallel computation. If NULL, the function runs the bootstrap sequentially.}

\item{R}{The number of bootstrap resamples to generate.}
}
\value{
A nested dataframe with an additional column bootstrap containing the bootstrapped results.
}
\description{
This function performs bootstrapping on a nested dataframe with a nested column called statistics.
}
