% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sample.R
\name{sample_methods}
\alias{sample_methods}
\title{Apply sample methods}
\usage{
sample_methods(data, nSamp, iter, method)
}
\arguments{
\item{data}{A data frame to be sampled.}

\item{nSamp}{The number of samples to generate.}

\item{iter}{The number of iterations to perform.}

\item{method}{The method to use for sampling. Can be "lhs" for Latin Hypercube Sampling,
"srs" for Simple Random Sampling, or "lpm" else for Balanced Sampling.}
}
\description{
Apply sample methods
}
