% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples}
\alias{samples}
\title{Monte Carlo samples}
\format{
\code{sf} object.
}
\usage{
samples
}
\description{
This is an \code{tibble} of monte carlo samples from the \code{monte_carlo()} function.
}
\keyword{datasets}
