% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stats}
\alias{stats}
\title{Statistics of samples}
\format{
\code{sf} object.
}
\usage{
stats
}
\description{
This is a nested \code{tibble} of statistics for \code{samples}
}
\keyword{datasets}
