% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{stats_nested}
\alias{stats_nested}
\title{Calculate nested summary statistics}
\usage{
stats_nested(data, metrics = NULL, cores = NULL, .f = NULL)
}
\arguments{
\item{data}{A data frame containing the data to be summarized}

\item{metrics}{A vector of metric names to be summarized. If \code{NULL}, all numeric columns will be summarized.}

\item{cores}{Number of cores to use for parallel computation}
}
\value{
A tibble containing the nested data frames and their corresponding summary statistics.
}
\description{
This function takes a data frame and creates a nested data frame by grouping on the columns \code{nSamp}, \code{iter}, and \code{method}.
It then calculates summary statistics for each nested data frame using the \code{summary_stats} function.
}
\examples{
data("samples")
stats_nested(samples, metrics = c("zmean", "zq90"))

}
