% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_plot.R
\name{stats_plot}
\alias{stats_plot}
\title{Plot summary statistics}
\usage{
stats_plot(data, statistics = NULL, type = "mean", ...)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{statistics}{An optional vector of statistics to include in the plot.}

\item{type}{The type of plot to create, either "mean" (default) or "box".}

\item{...}{Additional arguments passed to ggplot2.}
}
\value{
A ggplot2 object representing the plot of summary statistics.
}
\description{
This function creates a plot of summary statistics using ggplot2.
}
