% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sample.R
\name{stdmethods}
\alias{stdmethods}
\title{Apply Sampling Method to Data}
\usage{
stdmethods(data, nSamp, iter, method)
}
\arguments{
\item{data}{A data frame to be sampled.}

\item{nSamp}{The number of samples to generate.}

\item{iter}{The number of iterations to perform.}

\item{method}{The method to use for sampling. Can be "lhs" for Latin Hypercube Sampling,
"srs" for Simple Random Sampling, or "lpm" else for Balanced Sampling.}
}
\value{
A data frame containing the sampled data, the iteration number, the number of samples, and the sampling method used.
}
\description{
This function applies one of three sampling methods to data: Latin Hypercube Sampling (LHS), Simple Random Sampling (SRS), or Latin Point Mass Sampling (LPM).
}
