% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{stdsummary}
\alias{stdsummary}
\title{Standard summary}
\usage{
stdsummary(x, population, R = 10000)
}
\arguments{
\item{x}{A numeric vector for which summary statistics are to be computed.}

\item{population}{dataframe with population statistics}
}
\description{
Standard summary
}
