% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_stats}
\alias{util_stats}
\title{Calculate descriptive statistics for each column in a data frame}
\usage{
util_stats(metrics, population = FALSE)
}
\arguments{
\item{metrics}{A data frame with numeric columns to calculate statistics for.}

\item{population}{Logical defining whether \code{metrics} is a population. If \code{population = FALSE}
then a standard error will be calculated.}
}
\value{
A data frame with columns for each calculated statistic, where each row represents a column in the input data frame.
}
\description{
Given a data frame, calculates the standard error, minimum, mean, maximum, variance, interquartile range (IQR), and decile quantiles for each column.
}
\examples{
data(iris)
util_stats(iris[, 1:4])
}
