% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{utils_list_to_vector}
\alias{utils_list_to_vector}
\title{Create a vector by repeating values from a list}
\usage{
utils_list_to_vector(nSamp, iter)
}
\arguments{
\item{nSamp}{Vector of desired sample sizes}

\item{iter}{Desired number of iterations}
}
\value{
A vector for each \code{nSamp} of length \code{iter}
}
\description{
This function takes a list of values and repeats each value a specified number of times to create a vector.
}
