% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_sample_applied}
\alias{utils_sample_applied}
\title{Sampling of Spatial Data}
\usage{
utils_sample_applied(data, nSamp, iter, method)
}
\arguments{
\item{data}{A spatial dataset of class sf, SpatialPointsDataFrame, SpatialPixelsDataFrame or SpatialGridDataFrame.}

\item{nSamp}{The number of samples to take.}

\item{iter}{The iteration number.}

\item{method}{The method to use for sampling. Can be "lhs" for Latin Hypercube Sampling, "srs" for Simple Random Sampling, or anything else for Balanced Sampling.}
}
\value{
A data.frame object containing the sampled points and their attributes, as well as additional columns specifying the iteration number, the number of samples, and the sampling method used.
}
\description{
This function samples \code{nSamp}points \code{iter} times using either Latin Hypercube Sampling (LHS), Simple Random Sampling (SRS), or Balanced Sampling.
}
