% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_sample_bal}
\alias{utils_sample_bal}
\title{Sample balanced subsets of a large dataset}
\usage{
utils_sample_bal(data, nSamp, p = NULL)
}
\arguments{
\item{data}{A data.frame containing the dataset to be sampled}

\item{nSamp}{The number of samples to be taken}

\item{p}{A vector of probabilities to sample each row. If not provided, it defaults to uniform sampling.}
}
\value{
A data.frame containing the sampled rows
}
\description{
This function samples a balanced subset of a large dataset using the LPM2 method from the SamplingBigData package.
The method works by dividing the data into cells, and sampling from each cell in proportion to its size.
}
