% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahelsmethod.R
\name{ahelsmethod}
\alias{ahelsmethod}
\title{Sample ahels using the 'sample_ahels' function from the sgsR package.}
\usage{
ahelsmethod(existing, nFrac, mraster, ...)
}
\arguments{
\item{existing}{A data frame of existing samples}

\item{nFrac}{Scalar to use to define the number of samples to add using ahels}

\item{mraster}{A raster object acting as the population}

\item{...}{Further arguments passed to the \code{\link{ahelsmethod}}
function.}
}
\value{
A dataframe containing the sampled ahels.
}
\description{
Sample ahels using the 'sample_ahels' function from the sgsR package.
}
