% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sample.R
\name{apply_methods}
\alias{apply_methods}
\title{Sample a data frame using different methods}
\usage{
apply_methods(data, nSamp, iter, method = NULL, cores = NULL)
}
\arguments{
\item{data}{A data frame to be sampled.}

\item{nSamp}{The number of samples to generate.}

\item{iter}{The number of iterations to perform.}

\item{method}{The method to use for sampling. Can be "lhs" for Latin Hypercube Sampling,
"srs" for Simple Random Sampling, or "lpm" else for Balanced Sampling.}

\item{cores}{Number of cores to use for parallel computation}
}
\value{
A data frame containing the sampled data using each of the three methods.
}
\description{
This function takes a data frame and performs sampling using three different methods:
latin hypercube sampling (LHS), simple random sampling (SRS),
and Latin pseudo-maximum sampling (LPM).
}
\keyword{internal}
