% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-boot.R
\name{boot_mean}
\alias{boot_mean}
\title{Bootstrapped mean}
\usage{
boot_mean(x, i)
}
\arguments{
\item{x}{A vector of numerical values.}

\item{i}{A vector of indices specifying the current bootstrap sample.}
}
\value{
The bootstrapped mean of a sample.
}
\description{
This function computes the bootstrapped mean of a given sample.
}
\keyword{internal}
