% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-boot.R
\name{boot_median}
\alias{boot_median}
\title{Bootstrapped difference from population median}
\usage{
boot_median(x, i, pop)
}
\arguments{
\item{x}{A vector of numerical values.}

\item{i}{A vector of indices specifying the current bootstrap sample.}

\item{pop}{The population median to compare to.}
}
\value{
The difference between the median of the current bootstrap sample and the population median.
}
\description{
This function computes the difference between the median of a given sample and the population median using bootstrapping.
}
