% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plots}
\alias{plots}
\title{Plots data}
\format{
\code{sf} object.
}
\usage{
plots
}
\description{
This is an \code{sf} object containing the spatial data for 1000 plots with ALS metrics.
}
\keyword{datasets}
