% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sample.R
\name{sample_balanced}
\alias{sample_balanced}
\title{Sample balanced subsets of a large dataset}
\usage{
sample_balanced(data, nSamp, p = NULL)
}
\arguments{
\item{data}{A data frame to be sampled.}

\item{nSamp}{The number of samples to generate.}

\item{p}{A vector of probabilities to sample each row. If not provided, it defaults to uniform sampling.}
}
\value{
A data.frame containing the sampled rows
}
\description{
This function samples a balanced subset of a large dataset using the LPM2 method from the SamplingBigData package.
The method works by dividing the data into cells, and sampling from each cell in proportion to its size.
}
