% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-samples.R
\docType{data}
\name{samples}
\alias{samples}
\title{Samples data}
\format{
\code{sf} object.
}
\source{
This data was collected during a study of forest plots.
}
\usage{
samples
}
\description{
This is an \code{sf} object containing the spatial data for Monte Carlo samples in a study area.
}
\keyword{datasets}
