% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples}
\alias{samples}
\title{Samples from monte_carlo()}
\format{
\code{sf} object.
}
\usage{
samples
}
\description{
This is an \code{sf} object containing subsampled plots.
}
\keyword{datasets}
