% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stats}
\alias{stats}
\title{Statistics from \code{samples}}
\format{
\code{sf} object.
}
\usage{
stats
}
\description{
This is a nested dataframe object containing the statistics dervied from \code{samples}.
}
\keyword{datasets}
