% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_summary.R
\name{stats_summary}
\alias{stats_summary}
\title{Compute Summary Statistics}
\usage{
stats_summary(data, metrics = NULL, population = FALSE)
}
\arguments{
\item{data}{A data frame containing the data to be summarized}

\item{metrics}{A vector of metric names to be summarized. If \code{NULL}, all numeric columns will be summarized.}

\item{population}{A logical indicating whether the summary statistics should be calculated for the population or just the sample. Default is \code{FALSE}.}
}
\value{
A tibble object containing the computed summary statistics.
}
\description{
This function computes summary statistics for numeric columns in a SpatRaster or a data.frame. If a list of metrics is specified, it only computes summary statistics for those columns.
}
\examples{
data("samples")

# compute summary statistics for all numeric columns
stats_summary(samples)

# compute summary statistics for specific columns
stats_summary(samples, c("zmean", "zq90"))

}
