% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-boot.R
\name{stdboot}
\alias{stdboot}
\title{Standard summary}
\usage{
stdboot(x, R = 10000)
}
\arguments{
\item{x}{A numeric vector for which summary statistics are to be computed.}

\item{R}{Number of bootstrap iterations}

\item{population}{dataframe with population statistics}
}
\value{
Tibble with the bootstrapped statistic, bias, std.error,
conf.low, conf/high, bootstrap statistic, whether no varion resulting in an error '1 or 0',
and the bootstrap data 't'.
}
\description{
Standard summary
}
\keyword{internal}
