% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-boot.R
\name{stdboot}
\alias{stdboot}
\title{Standard summary}
\usage{
stdboot(x, population, R = 10000)
}
\arguments{
\item{x}{A numeric vector for which summary statistics are to be computed.}

\item{population}{dataframe with population statistics}

\item{R}{Number of bootstrap iterations}
}
\description{
Standard summary
}
